requirements.txt
# Everything needed in our project

# Micro WEB development framework
Flask~=2.0.3

# WEB Server just like nginx
gunicorn

# Python WEB request library
requests~=2.27.1

# Localization
flask-babel

# Database helper
Flask-SQLAlchemy~=2.0

# Database migration helper
flask-migrate
alembic~=1.7.6

# Database connection driver
SQLAlchemy<1.4
pymysql

# Environment variables with dotenv
python-dotenv~=0.19.2

# Admin panel>1.6.0
flask-admin

# User session management
flask-login

# Password secure
werkzeug~=2.0.3

# Swagger implementation for Flask
flasgger~=0.9.5
PyJWT~=1.7.1